Move all receiver code into receive component
