+++ src/NServiceBus.Core/ReceiveComponent.cs
@@ -2,6 +2,7 @@ namespace NServiceBus
 {
     using System;
     using System.Collections.Generic;
+    using System.Linq;
     using System.Threading.Tasks;
     using ConsistencyGuarantees;
     using Logging;
@@ -57,11 +58,58 @@ namespace NServiceBus
             receiveInfrastructure = transportInfrastructure.ConfigureReceiveInfrastructure();
         }
 
-        public List<TransportReceiver> CreateReceivers(MainPipelineExecutor mainPipelineExecutor, IEventAggregator eventAggregator, IBuilder builder, CriticalError criticalError)
+        public async Task InitializeReceivers(MainPipelineExecutor mainPipelineExecutor, IEventAggregator eventAggregator, IBuilder builder, CriticalError criticalError)
+        {
+            AddReceivers(mainPipelineExecutor, eventAggregator, builder, criticalError);
+
+            foreach (var receiver in receivers)
+            {
+                try
+                {
+                    await receiver.Init().ConfigureAwait(false);
+                }
+                catch (Exception ex)
+                {
+                    Logger.Fatal($"Receiver {receiver.Id} failed to initialize.", ex);
+                    throw;
+                }
+            }
+        }
+
+        public void Start()
+        {
+            foreach (var receiver in receivers)
+            {
+                try
+                {
+                    receiver.Start();
+                }
+                catch (Exception ex)
+                {
+                    Logger.Fatal($"Receiver {receiver.Id} failed to start.", ex);
+                    throw;
+                }
+            }
+        }
+
+        public Task Stop()
+        {
+            var receiverStopTasks = receivers.Select(async receiver =>
+            {
+                Logger.DebugFormat("Stopping {0} receiver", receiver.Id);
+                await receiver.Stop().ConfigureAwait(false);
+                Logger.DebugFormat("Stopped {0} receiver", receiver.Id);
+            });
+
+            return Task.WhenAll(receiverStopTasks);
+        }
+
+
+        void AddReceivers(MainPipelineExecutor mainPipelineExecutor, IEventAggregator eventAggregator, IBuilder builder, CriticalError criticalError)
         {
             if (isSendOnlyEndpoint)
             {
-                return new List<TransportReceiver>();
+                return;
             }
 
             var purgeOnStartup = settings.GetOrDefault<bool>("Transport.PurgeOnStartup");
@@ -78,7 +126,6 @@ namespace NServiceBus
             var pushSettings = new PushSettings(LocalAddress, errorQueue, purgeOnStartup, requiredTransactionSupport);
             var dequeueLimitations = GetDequeueLimitationsForReceivePipeline();
 
-            var receivers = new List<TransportReceiver>();
             receivers.Add(new TransportReceiver(MainReceiverId, BuildMessagePump(), pushSettings, dequeueLimitations, mainPipelineExecutor, recoverabilityExecutor, criticalError));
 
             if (InstanceSpecificQueue != null)
@@ -97,25 +144,6 @@ namespace NServiceBus
 
                 receivers.Add(new TransportReceiver(satellitePipeline.Name, BuildMessagePump(), satellitePushSettings, satellitePipeline.RuntimeSettings, new SatellitePipelineExecutor(builder, satellitePipeline), satelliteRecoverabilityExecutor, criticalError));
             }
-
-            return receivers;
-        }
-
-        //note: this should be handled in a feature but we don't have a good
-        // extension point to plugin atm
-        PushRuntimeSettings GetDequeueLimitationsForReceivePipeline()
-        {
-            if (settings.TryGet(out MessageProcessingOptimizationExtensions.ConcurrencyLimit concurrencyLimit))
-            {
-                return new PushRuntimeSettings(concurrencyLimit.MaxValue);
-            }
-
-            return PushRuntimeSettings.Default;
-        }
-
-        IPushMessages BuildMessagePump()
-        {
-            return receiveInfrastructure.MessagePumpFactory();
         }
 
         public Task CreateQueuesIfNecessary(string username)
@@ -151,6 +179,25 @@ namespace NServiceBus
             }
         }
 
+        //note: this should be handled in a feature but we don't have a good
+        // extension point to plugin atm
+        PushRuntimeSettings GetDequeueLimitationsForReceivePipeline()
+        {
+            if (settings.TryGet(out MessageProcessingOptimizationExtensions.ConcurrencyLimit concurrencyLimit))
+            {
+                return new PushRuntimeSettings(concurrencyLimit.MaxValue);
+            }
+
+            return PushRuntimeSettings.Default;
+        }
+
+        IPushMessages BuildMessagePump()
+        {
+            return receiveInfrastructure.MessagePumpFactory();
+        }
+
+
+        List<TransportReceiver> receivers = new List<TransportReceiver>();
         TransportReceiveInfrastructure receiveInfrastructure;
         TransportInfrastructure transportInfrastructure;
