+++ src/NServiceBus.Core/Unicast/RunningEndpointInstance.cs
@@ -13,11 +13,11 @@ namespace NServiceBus
 
     class RunningEndpointInstance : IEndpointInstance
     {
-        public RunningEndpointInstance(SettingsHolder settings, IBuilder builder, List<TransportReceiver> receivers, FeatureRunner featureRunner, IMessageSession messageSession, TransportInfrastructure transportInfrastructure)
+        public RunningEndpointInstance(SettingsHolder settings, IBuilder builder, ReceiveComponent receiving, FeatureRunner featureRunner, IMessageSession messageSession, TransportInfrastructure transportInfrastructure)
         {
             this.settings = settings;
             this.builder = builder;
-            this.receivers = receivers;
+            this.receiving = receiving;
             this.featureRunner = featureRunner;
             this.messageSession = messageSession;
             this.transportInfrastructure = transportInfrastructure;
@@ -42,7 +42,7 @@ namespace NServiceBus
                 Log.Info("Initiating shutdown.");
 
                 // Cannot throw by design
-                await StopReceivers().ConfigureAwait(false);
+                await receiving.Stop().ConfigureAwait(false);
                 await featureRunner.Stop(messageSession).ConfigureAwait(false);
                 // Can throw
                 await transportInfrastructure.Stop().ConfigureAwait(false);
@@ -93,20 +93,8 @@ namespace NServiceBus
             return messageSession.Unsubscribe(eventType, options);
         }
 
-        Task StopReceivers()
-        {
-            var receiverStopTasks = receivers.Select(async receiver =>
-            {
-                Log.DebugFormat("Stopping {0} receiver", receiver.Id);
-                await receiver.Stop().ConfigureAwait(false);
-                Log.DebugFormat("Stopped {0} receiver", receiver.Id);
-            });
-
-            return Task.WhenAll(receiverStopTasks);
-        }
-
         IBuilder builder;
-        List<TransportReceiver> receivers;
+        ReceiveComponent receiving;
         FeatureRunner featureRunner;
         IMessageSession messageSession;
         TransportInfrastructure transportInfrastructure;
